requirements.txt
accessify==0.3.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.6
asttokens==2.0.8
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autovizwidget==0.20.0
azure-common==1.1.28
azure-core==1.29.3
azure-identity==1.14.0
azure-mgmt-core==1.4.0
azure-mgmt-datafactory==3.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.1.0
fastjsonschema==2.18.0
filelock==3.8.0
fqdn==1.5.1
frozenlist==1.4.0
hdijupyterutils==0.20.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
isoduration==20.11.0
isort==5.12.0
jedi==0.18.1
jep==4.1.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==7.3.5
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.24.0
lazy-object-proxy==1.9.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mock==4.0.3
msal==1.23.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.9.2
nest-asyncio==1.5.5
nose==1.3.7
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.23.3
oauthlib==3.2.2
openai==0.27.9
overrides==7.4.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.10.0
pluggy==1.2.0
portalocker==2.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.8.0
pylint==2.17.5
pyparsing==3.0.9
pyrsistent==0.18.1
pyspnego==0.6.0
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.2.1
pywin32==306
pywinpty==2.0.8
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-kerberos==0.14.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sparkmagic==0.13.1
stack-data==0.5.1
tenacity==8.1.0
terminado==0.15.0
tiktoken==0.4.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.16.5
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
widgetsnbextension==4.0.3
wrapt==1.15.0
yarl==1.9.2
